\name{std}
\alias{std}

\title{An example function from the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R}

\description{An example function from Chapter 17 of the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. See Appendix B or http://www.wiley.com/go/data_mining_algorithms for more details.}

\usage{See Section 17.3, Example 17.3.1.}

\arguments{See Section 17.3, Example 17.3.1.}

\details{See Section 17.3, Example 17.3.1.}

\value{See Section 17.3, Example 17.3.1.}

\references{Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. Wiley.}

\author{
Pawel Cichosz <p.cichosz@elka.pw.edu.pl>
}

\note{
}


\seealso{
\code{std.all}
\code{nrm}
}

\examples{
data(weatherc, package="dmr.data")

std(weatherc$temperature)
}

\keyword{arith}
